* Replaced shorts with bytes and int8_t; Significantly reduces code size, and will most likely reduce RAM usage and increase execution speed. * Lowered RAM usage of Interrupt Chain. * Some code comment fixes.

+++ InterruptChain/examples/ReceiveRemoteAndSensor/ReceiveRemoteAndSensor.ino
@@ -95,7 +95,7 @@ void showTempHumi(byte *data) {
     // Yes!
     byte channel, randomId;
     int temp;
-    short int humidity;
+    byte humidity;
 
     // Decode the data
     SensorReceiver::decodeThermoHygro(data, channel, randomId, temp, humidity);
